[html-elements] Allow non-text children for <A>
Currently, rendering a view inside of <A /> shows nothing on native. I ran into a similar problem with [expo-next-react-navigation](https://github.com/nandorojo/expo-next-react-navigation/issues/13), and fixed it with an `isText` prop, which defaults to true. 

I'm proposing a change similar to this here. If you have a better solution, that works too. For now, I'll have to recreate this component in my own app.

Thanks!