Update customizing-webpack.md
This PR is a roundabout way of making an issue – the `ResizeObserver` polyfill injection doesn't seem to always be working. On iOS 13.3 Safari devices, if I don't include this at the top, then `onLayout` is never called. I can send a repro later and turn this into an issue if needed. I don't have time to repro it at the moment since I need to push this out for users that are seeing a blank screen, but I will ASAP.

Technically, this PR shouldn't be needed, since the intent is that the polyfill is injected by expo webpack automatically. I'll investigate further to see if that's a bug with webpack or my configuration of next.config.js.